refactor: extract kernel syscalls module
