+++ core/kernel/process.ts
@@ -1,7 +1,7 @@
 // Process management utilities for the Helios-OS Kernel
 
 import { invoke } from '@tauri-apps/api/tauri';
-import type { SyscallDispatcher } from './index';
+import type { SyscallDispatcher } from './syscalls';
 import type { Kernel } from './index';
 
 export type ProcessID = number;
